Makefile.am: link with libiconv when needed
Fixes issues on systems with libiconv separate from libc
(e.g. Darwin/macOS, uclibc).

We already ask configure to do detection for us
of whether we need an explicit library link line
for libintl and libiconv, so let's use the result.

This will have no effect if the variable is empty
of course, as will be the case on many systems
(everywhere it worked before).

We've been carrying this patch in Gentoo for quite some time
with no issues.

Bug: https://bugs.gentoo.org/366885
Bug: https://bugs.gentoo.org/499940
Signed-off-by: Sam James <sam@gentoo.org>